requirements.txt
pandas
python-dotenv
openai  # or anthropic
